Limit data to tenant
