+++ backend/src/api/integration/integrationGlobal.ts
@@ -4,7 +4,10 @@ import PermissionChecker from '../../services/user/permissionChecker'
 
 export default async (req, res) => {
   new PermissionChecker(req).validateHas(Permissions.values.integrationRead)
-  const payload = await new IntegrationService(req).findGlobalIntegrations(req.query)
+  const payload = await new IntegrationService(req).findGlobalIntegrations(
+    req.params.tenantId,
+    req.query,
+  )
 
   await req.responseHandler.success(req, res, payload)
 }
