+++ backend/src/api/integration/integrationGlobalStatus.ts
@@ -5,7 +5,10 @@ import PermissionChecker from '../../services/user/permissionChecker'
 export default async (req, res) => {
   new PermissionChecker(req).validateHas(Permissions.values.integrationRead)
 
-  const payload = await new IntegrationService(req).findGlobalIntegrationsStatusCount(req.query)
+  const payload = await new IntegrationService(req).findGlobalIntegrationsStatusCount(
+    req.params.tenantId,
+    req.query,
+  )
 
   await req.responseHandler.success(req, res, payload)
 }
