+++ backend/src/database/repositories/integrationRepository.ts
@@ -400,34 +400,61 @@ class IntegrationRepository {
     })
   }
 
+  /**
+   * Finds global integrations based on the provided parameters.
+   *
+   * @param {string} tenantId - The ID of the tenant for which integrations are to be found.
+   * @param {Object} filters - An object containing various filter options.
+   * @param {string} [filters.platform=null] - The platform to filter integrations by.
+   * @param {string[]} [filters.status=['done']] - The status of the integrations to be filtered.
+   * @param {string} [filters.query=''] - The search query to filter integrations.
+   * @param {number} [filters.limit=20] - The maximum number of integrations to return.
+   * @param {number} [filters.offset=0] - The offset for pagination.
+   * @param {IRepositoryOptions} options - The repository options for querying.
+   * @returns {Promise<Object>} The result containing the rows of integrations and metadata about the query.
+   */
   static async findGlobalIntegrations(
+    tenantId: string,
     { platform = null, status = ['done'], query = '', limit = 20, offset = 0 },
     options: IRepositoryOptions,
   ) {
     const qx = SequelizeRepository.getQueryExecutor(options)
     if (status.includes('not-connected')) {
-      const rows = await fetchGlobalNotConnectedIntegrations(qx, platform, query, limit, offset)
-      const [result] = await fetchGlobalNotConnectedIntegrationsCount(qx, platform, query)
+      const rows = await fetchGlobalNotConnectedIntegrations(
+        qx,
+        tenantId,
+        platform,
+        query,
+        limit,
+        offset,
+      )
+      const [result] = await fetchGlobalNotConnectedIntegrationsCount(qx, tenantId, platform, query)
       return { rows, count: +result.count, limit: +limit, offset: +offset }
     }
 
-    const rows = await fetchGlobalIntegrations(qx, status, platform, query, limit, offset)
-    const [result] = await fetchGlobalIntegrationsCount(qx, status, platform, query)
+    const rows = await fetchGlobalIntegrations(qx, tenantId, status, platform, query, limit, offset)
+    const [result] = await fetchGlobalIntegrationsCount(qx, tenantId, status, platform, query)
     return { rows, count: +result.count, limit: +limit, offset: +offset }
   }
 
   /**
-   * Retrieves the count of global integration statuses, including a 'not-connected' status.
+   * Retrieves the count of global integrations statuses for a specified tenant and platform.
+   * This method aggregates the count of different integration statuses including a 'not-connected' status.
    *
-   * @param {Object} param0 - An object containing the parameters.
-   * @param {string|null} param0.platform - The platform for which to fetch the integration statuses. Default is null.
-   * @param {IRepositoryOptions} options - Options for the repository including the query executor.
-   * @return {Promise<Array<Object>>} A promise that resolves to an array of objects, each representing an integration status with a count. Each object contains a 'status' and a 'count' property.
+   * @param {string} tenantId - The unique identifier for the tenant.
+   * @param {Object} param1 - The optional parameters.
+   * @param {string|null} [param1.platform=null] - The platform to filter the integrations. Default is null.
+   * @param {IRepositoryOptions} options - The options for the repository operations.
+   * @return {Promise<Array<Object>>} A promise that resolves to an array of objects containing the statuses and their counts.
    */
-  static async findGlobalIntegrationsStatusCount({ platform = null }, options: IRepositoryOptions) {
+  static async findGlobalIntegrationsStatusCount(
+    tenantId: string,
+    { platform = null },
+    options: IRepositoryOptions,
+  ) {
     const qx = SequelizeRepository.getQueryExecutor(options)
-    const [result] = await fetchGlobalNotConnectedIntegrationsCount(qx, platform, '')
-    const rows = await fetchGlobalIntegrationsStatusCount(qx, platform)
+    const [result] = await fetchGlobalNotConnectedIntegrationsCount(qx, tenantId, platform, '')
+    const rows = await fetchGlobalIntegrationsStatusCount(qx, tenantId, platform)
     return [...rows, { status: 'not-connected', count: +result.count }]
   }
