+++ services/libs/data-access-layer/src/integrations/index.ts
@@ -3,18 +3,20 @@ import { IIntegration } from '@crowd/types'
 import { QueryExecutor } from '../queryExecutor'
 
 /**
- * Fetches global integrations based on the provided filters and pagination parameters.
+ * Fetches a list of global integrations based on the provided filters.
  *
- * @param {QueryExecutor} qx - The query executor to run the database queries.
- * @param {string[]} status - An array of statuses to filter the integrations.
- * @param {string | null} platform - The specific platform to filter the integrations, or null to include all platforms.
- * @param {string} query - A search query to filter integrations by segment name.
- * @param {number} limit - The maximum number of integrations to return.
- * @param {number} offset - The number of integrations to skip before starting to collect the result set.
- * @return {Promise<IIntegration[]>} - A promise that resolves to an array of integrations matching the criteria.
+ * @param {QueryExecutor} qx - The query executor object to perform database queries.
+ * @param {string} tenantId - The ID of the tenant for which integrations are being fetched.
+ * @param {string[]} status - An array of status values to filter the integrations.
+ * @param {string | null} platform - The platform to filter integrations, or null to include all platforms.
+ * @param {string} query - A search string to filter segment names.
+ * @param {number} limit - The maximum number of results to return.
+ * @param {number} offset - The number of results to skip before starting to collect the result set.
+ * @return {Promise<IIntegration[]>} A promise that resolves to the list of integrations matching the filters.
  */
 export async function fetchGlobalIntegrations(
   qx: QueryExecutor,
+  tenantId: string,
   status: string[],
   platform: string | null,
   query: string,
@@ -37,6 +39,7 @@ export async function fetchGlobalIntegrations(
                JOIN segments s ON i."segmentId" = s.id
         WHERE i."status" = ANY ($(status)::text[])
           AND i."deletedAt" IS NULL
+          AND i."tenantId" = $(tenantId)
           AND ($(platform) IS NULL OR i."platform" = $(platform))
           AND s.name ILIKE $(query)
         LIMIT $(limit) OFFSET $(offset)
@@ -44,6 +47,7 @@ export async function fetchGlobalIntegrations(
     {
       status,
       platform,
+      tenantId,
       query: `%${query}%`,
       limit,
       offset,
@@ -52,16 +56,18 @@ export async function fetchGlobalIntegrations(
 }
 
 /**
- * Fetches the count of global integrations based on specified criteria.
+ * Fetches the count of global integrations based on the specified criteria.
  *
- * @param {QueryExecutor} qx - The query executor instance used to run the SQL query.
- * @param {string[]} status - An array of status strings to filter the integrations.
- * @param {string | null} platform - The platform to filter the integrations, or null for no platform filter.
- * @param {string} query - The query string to filter the segment names.
- * @return {Promise<{ count: number }[]>} A promise that resolves to an array of objects containing the count of integrations.
+ * @param {QueryExecutor} qx - The query executor to run the database query.
+ * @param {string} tenantId - The tenant identifier to filter integrations.
+ * @param {string[]} status - The array of statuses to filter integrations.
+ * @param {string|null} platform - The platform to filter by, or null for all platforms.
+ * @param {string} query - The query string to filter segment names.
+ * @return {Promise<{ count: number }[]>} The promise that resolves to an array with the count of integrations.
  */
 export async function fetchGlobalIntegrationsCount(
   qx: QueryExecutor,
+  tenantId: string,
   status: string[],
   platform: string | null,
   query: string,
@@ -73,30 +79,34 @@ export async function fetchGlobalIntegrationsCount(
                JOIN segments s ON i."segmentId" = s.id
         WHERE i."status" = ANY ($(status)::text[])
           AND i."deletedAt" IS NULL
+          AND i."tenantId" = $(tenantId)
           AND ($(platform) IS NULL OR i."platform" = $(platform))
           AND s.name ILIKE $(query)
       `,
     {
       status,
       platform,
+      tenantId,
       query: `%${query}%`,
     },
   )
 }
 
 /**
- * Fetches the list of globally not connected integrations based on provided criteria such as platform, query, limit, and offset.
+ * Fetches a list of global integrations that are not connected.
  *
- * @param {QueryExecutor} qx - The query executor instance to run the SQL queries.
- * @param {string | null} platform - The platform to filter the integrations, can be null to include all platforms.
- * @param {string} query - The search query to filter the integration names.
- * @param {number} limit - The number of results to return.
- * @param {number} offset - The offset for pagination.
+ * @param {QueryExecutor} qx - The query executor to run the queries.
+ * @param {string} tenantId - The tenant ID to filter the integrations.
+ * @param {string | null} platform - The specific platform to filter the integrations, or null for all platforms.
+ * @param {string} query - The query string to filter by integration name.
+ * @param {number} limit - The maximum number of integrations to return.
+ * @param {number} offset - The number of integrations to skip before starting to collect the result set.
  *
- * @return {Promise<IIntegration[]>} A promise that resolves to an array of not connected integrations.
+ * @return {Promise<IIntegration[]>} A promise that resolves to an array of integrations not connected to the specified platform.
  */
 export async function fetchGlobalNotConnectedIntegrations(
   qx: QueryExecutor,
+  tenantId: string,
   platform: string | null,
   query: string,
   limit: number,
@@ -124,12 +134,14 @@ export async function fetchGlobalNotConnectedIntegrations(
       WHERE cp.platform IS NULL
         AND s."parentId" IS NOT NULL
         AND s."grandparentId" IS NOT NULL
+        AND s."tenantId" = $(tenantId)
         AND ($(platform) IS NULL OR up."platform" = $(platform))
         AND s.name ILIKE $(query)
       LIMIT $(limit) OFFSET $(offset)
     `,
     {
       platform,
+      tenantId,
       query: `%${query}%`,
       limit,
       offset,
@@ -138,15 +150,17 @@ export async function fetchGlobalNotConnectedIntegrations(
 }
 
 /**
- * Fetches the count of globally not connected integrations based on the provided criteria.
+ * Fetches the count of global integrations that are not connected.
  *
- * @param {QueryExecutor} qx - The query executor instance used to run database queries.
- * @param {string | null} platform - The platform name to filter the integrations (nullable).
- * @param {string} query - The search query to filter segments by name.
- * @return {Promise<{ count: number }[]>} A promise that resolves to an array of objects containing the count of not connected integrations.
+ * @param {QueryExecutor} qx - The query executor used to perform SQL queries.
+ * @param {string} tenantId - The ID of the tenant for whom integrations need to be fetched.
+ * @param {string|null} platform - The platform to filter results by (optional).
+ * @param {string} query - The name pattern to filter segments by.
+ * @return {Promise<{count: number}[]>} - A promise that resolves to an array of objects containing the count of not connected integrations.
  */
 export async function fetchGlobalNotConnectedIntegrationsCount(
   qx: QueryExecutor,
+  tenantId: string,
   platform: string | null,
   query: string,
 ): Promise<{ count: number }[]> {
@@ -166,25 +180,29 @@ export async function fetchGlobalNotConnectedIntegrationsCount(
       WHERE cp.platform IS NULL
         AND s."parentId" IS NOT NULL
         AND s."grandparentId" IS NOT NULL
+        AND s."tenantId" = $(tenantId)
         AND ($(platform) IS NULL OR up."platform" = $(platform))
         AND s.name ILIKE $(query)
     `,
     {
       platform,
+      tenantId,
       query: `%${query}%`,
     },
   )
 }
 
 /**
- * Fetches the count of global integrations statuses from the database.
+ * Fetches the count of integrations grouped by their status for a given tenant and optional platform.
  *
- * @param {QueryExecutor} qx - Query executor to run the database queries.
- * @param {string|null} platform - The platform to filter integrations by. If null, no platform filtering is applied.
- * @return {Promise<{status: string, count: number}[]>} A promise that resolves to an array of objects containing integration status and their respective counts.
+ * @param {QueryExecutor} qx - The query executor used to perform the database query.
+ * @param {string} tenantId - The ID of the tenant for which to fetch the integrations status count.
+ * @param {string | null} platform - The platform to filter the integrations by, or null if no platform filter is to be applied.
+ * @return {Promise<{status: string, count: number}[]>} A promise that resolves to an array of objects, each containing a status and the corresponding count of integrations.
  */
 export async function fetchGlobalIntegrationsStatusCount(
   qx: QueryExecutor,
+  tenantId: string,
   platform: string | null,
 ): Promise<{ status: string; count: number }[]> {
   return qx.select(
@@ -193,11 +211,13 @@ export async function fetchGlobalIntegrationsStatusCount(
                COUNT(*) AS count
         FROM "integrations" i
         WHERE i."deletedAt" IS NULL
+          AND i."tenantId" = $(tenantId)
           AND ($(platform) IS NULL OR i."platform" = $(platform))
         GROUP BY i.status
     `,
     {
       platform,
+      tenantId,
     },
   )
 }
